+++ Backend/package.json
@@ -17,6 +17,7 @@
     "@adminjs/adonis": "^1.1.0",
     "@adminjs/express": "^6.1.0",
     "@adminjs/mongoose": "^4.1.0",
+    "@sendgrid/mail": "^8.1.5",
     "adminjs": "^7.8.13",
     "axios": "^1.7.8",
     "bcrypt": "^5.1.1",
